cocogpt-app-main/requirements.txt
trubrics==1.6.2
azure-core==1.29.5
azure-cosmos==4.5.1

ipython==8.12.3

langchain==0.0.240
langcodes==3.3.0
langsmith==0.0.66

loguru==0.7.2
openai==0.27.8
openapi-schema-pydantic==1.2.4

pydantic==1.10.13
pyodbc==5.0.1

requests==2.31.0

streamlit==1.28.2
streamlit-authenticator==0.2.3
streamlit-feedback==0.1.2


tiktoken==0.5.1


